refactor(x/erc20): separate concerns for erc20 mint, burn and transfer ownership operations
