+++ x/erc20/keeper/mint.go
@@ -111,96 +111,3 @@ func (k Keeper) MintCoins(ctx sdk.Context, sender, to sdk.AccAddress, amount mat
 
 	return nil
 }
-
-// BurnCoins burns the provided amount of coins from the given address.
-func (k Keeper) BurnCoins(ctx sdk.Context, sender sdk.AccAddress, amount math.Int, token string) error {
-	pair, found := k.GetTokenPair(ctx, k.GetTokenPairID(ctx, token))
-	if !found {
-		return errorsmod.Wrapf(types.ErrTokenPairNotFound, "token '%s' not registered", token)
-	}
-
-	if !pair.IsNativeCoin() {
-		return errorsmod.Wrap(types.ErrNonNativeCoinBurningDisabled, token)
-	}
-
-	coins := sdk.Coins{{Denom: pair.Denom, Amount: amount}}
-
-	err := k.bankKeeper.SendCoinsFromAccountToModule(ctx, sender, types.ModuleName, coins)
-	if err != nil {
-		return err
-	}
-
-	err = k.bankKeeper.BurnCoins(ctx, types.ModuleName, coins)
-	if err != nil {
-		return err
-	}
-
-	ctx.EventManager().EmitEvent(
-		sdk.NewEvent(
-			sdk.EventTypeMessage,
-			sdk.NewAttribute(sdk.AttributeKeyAction, types.TypeMsgBurn),
-			sdk.NewAttribute(sdk.AttributeKeyModule, types.ModuleName),
-			sdk.NewAttribute(sdk.AttributeKeySender, sender.String()),
-			sdk.NewAttribute(sdk.AttributeKeyAmount, amount.String()),
-		),
-	)
-	return nil
-}
-
-// TransferOwnershipProposal transfers ownership of the token to the new owner through a proposal
-func (k Keeper) TransferOwnershipProposal(ctx sdk.Context, newOwner sdk.AccAddress, token string) error {
-	pair, found := k.GetTokenPair(ctx, k.GetTokenPairID(ctx, token))
-	if !found {
-		return errorsmod.Wrapf(types.ErrTokenPairNotFound, "token '%s' not registered", token)
-	}
-
-	return k.transferOwnership(ctx, newOwner, pair)
-}
-
-// TransferOwnership transfers ownership of the token to the new owner.
-func (k Keeper) TransferOwnership(ctx sdk.Context, sender, newOwner sdk.AccAddress, token string) error {
-	pair, found := k.GetTokenPair(ctx, k.GetTokenPairID(ctx, token))
-	if !found {
-		return errorsmod.Wrapf(types.ErrTokenPairNotFound, "token '%s' not registered", token)
-	}
-
-	ownerAddr, err := sdk.AccAddressFromBech32(pair.OwnerAddress)
-	if err != nil {
-		return errorsmod.Wrapf(err, "invalid owner address")
-	}
-
-	if !sender.Equals(ownerAddr) {
-		return errorsmod.Wrap(types.ErrMinterIsNotOwner, "sender is not the owner of the token")
-	}
-
-	return k.transferOwnership(ctx, newOwner, pair)
-}
-
-// transferOwnership transfers ownership of the token to the new owner
-func (k Keeper) transferOwnership(ctx sdk.Context, newOwner sdk.AccAddress, token types.TokenPair) error {
-	if !token.IsNativeCoin() {
-		return errorsmod.Wrap(types.ErrNonNativeTransferOwnershipDisabled, token.Erc20Address)
-	}
-
-	k.SetTokenPairOwnerAddress(ctx, token, newOwner.String())
-
-	ctx.EventManager().EmitEvent(
-		sdk.NewEvent(
-			sdk.EventTypeMessage,
-			sdk.NewAttribute(sdk.AttributeKeyAction, types.TypeMsgTransferOwnership),
-			sdk.NewAttribute(sdk.AttributeKeyModule, types.ModuleName),
-			sdk.NewAttribute(types.AttributeKeyNewOwner, newOwner.String()),
-		),
-	)
-
-	return nil
-}
-
-func (k Keeper) GetOwnerAddress(ctx sdk.Context, contractAddress string) string {
-	pair, found := k.GetTokenPair(ctx, k.GetTokenPairID(ctx, contractAddress))
-	if !found {
-		return ""
-	}
-
-	return pair.OwnerAddress
-}
